project/requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.1